Change the for syntax to work on `IExpectSubject<T>` instead of `IThat<T>`
